Fix Travis warning: multi-character character constant [-Wmultichar]

+++ contrib/extractor/loadlib/loadlib.cpp
@@ -50,7 +50,7 @@ bool FileLoader::prepareLoadedData()
 {
     // Check version
     version = (file_MVER*) data;
-    if (version->fcc != 'MVER')
+    if (version->fcc != uint32('MVER'))
         return false;
     if (version->ver != FILE_FORMAT_VERSION)
         return false;
